Move VR Content from COMolecule into MoleculeVRInteractionHandler
