+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Scripts/Maroon.Experiments.Catalyst.asmdef
@@ -13,8 +13,7 @@
         "GUID:53500f10aff8aca4fab6cf9f9f64fcb1",
         "GUID:cd80b0e2ec900ac4cbf0ba07b69d4e3d",
         "GUID:dd8043639e4014317a7246f064330196",
-        "GUID:1fbf338ef6b952a448262a7ff7df5fe4",
-        "GUID:02377da00954557488f6d58db762644d"
+        "GUID:1fbf338ef6b952a448262a7ff7df5fe4"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
