--- unity/Assets/Maroon/scenes/experiments/Catalyst/Scripts/VR/Maroon.Experiments.Catalyst.VR.asmdef
@@ -1,3 +1,0 @@
-{
-	"name": "Maroon.Experiments.Catalyst.VR"
-}
