--- unity/Assets/Maroon/scenes/experiments/Catalyst/Scripts/VR/Maroon.Experiments.Catalyst.VR.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 02377da00954557488f6d58db762644d
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
