+++ unity/Assets/Maroon/scenes/experiments/Catalyst/Scripts/VR/MoleculeVRInteractionHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c77c800e82ca8c847b714150a07ad9a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
